requirements.txt
langchain
langgraph
langsmith
faiss-cpu
tiktoken
openai
anthropic
groq
gitpython
#tree-sitter
python-dotenv
streamlit
notebook
langchain-anthropic
#tree-sitter
#tree_sitter_languages
tree-sitter==0.24.0 
tree-sitter-languages==1.10.2
javalang

